Refactor event sourcing mocker into reusable parts
