+++ src/api/calculations/calculateSharePrice.js
@@ -1,0 +1,9 @@
+import { divide } from '~/utils/functionalBigNumber';
+import setPath from '~/api/utils/setPath';
+
+const calculateSharePrice = () =>
+  setPath(['calculations', 'sharePrice'], ({ data, calculations }) =>
+    divide(calculations.aum, calculations.totalSupply),
+  );
+
+export default calculateSharePrice;
