+++ src/api/modifications/addRedeem.js
@@ -1,0 +1,17 @@
+import { multiply } from '~/utils/functionalBigNumber';
+import setPath from '~/api/utils/setPath';
+
+const addRedeem = (amount, timestamp, investor) =>
+  setPath(['data', 'participations', 'list'], ({ data, calculations }) => [
+    ...data.participations.list,
+    {
+      investor,
+      token: data.meta.quoteToken,
+      type: 'redeem',
+      amount: multiply(calculations.sharePrice, amount),
+      shares: amount,
+      timestamp: timestamp || data.meta.inception,
+    },
+  ]);
+
+export default addRedeem;
