+++ src/api/queries/isSameToken.js
@@ -1,0 +1,3 @@
+const isSameToken = (a, b) => a.symbol === b.symbol && a.address === b.address;
+
+export default isSameToken;
