+++ src/api/utils/setPath.js
@@ -1,0 +1,6 @@
+import * as R from 'ramda';
+
+const setPath = (path, setter) => ({ data, calculations }) =>
+  R.assocPath(path, setter({ data, calculations }), { data, calculations });
+
+export default setPath;
